feat: add RM to landingpage
we use RM to tailor the pitch to the segmentation of the visitor
and when there is no pitch available from RM we default back to the
email signup.
add questions for the weekly newsletter to the subscribe paghe where I
tell the visitor to check their email and confirm.
I ask questions to determine which weekly newsletter they will get.
Adjust the visdual of the RM form on the start page and the way posts
are shown
fix: fix rendering of HTML inside of markdown (was disabled with Hugo
0.6